--- anomalib/models/utils/__init__.py
@@ -1,4 +1,0 @@
-"""Utility functions/classes needed by AnomalyModules."""
-from .blur import GaussianBlur2d
-
-__all__ = ["GaussianBlur2d"]
